Improve diagnostic messages in case of receptor signature warnings
